make toggle reducer global. move feature toggle query values into an array. add unit test
