+++ src/applications/facility-locator/reducers/index.js
@@ -1,9 +1,7 @@
 import { SearchResultReducer } from './searchResult';
 import { SearchQueryReducer } from './searchQuery';
-import { FeatureToggleReducer } from './featureToggle';
 
 const rootReducer = {
-  featureToggles: FeatureToggleReducer,
   searchResult: SearchResultReducer,
   searchQuery: SearchQueryReducer,
 };
