+++ src/applications/facility-locator/utils/actionTypes.js
@@ -3,8 +3,6 @@ export const FETCH_LOCATIONS = 'FETCH_LOCATIONS';
 export const FETCH_SERVICES = 'FETCH_SERVICES';
 export const FETCH_SERVICES_FAILED = 'FETCH_SERVICES_FAILED';
 export const FETCH_SERVICES_DONE = 'FETCH_SERVICES_DONE';
-export const FETCH_TOGGLE_VALUES_STARTED = 'FETCH_TOGGLE_VALUES_STARTED';
-export const FETCH_TOGGLE_VALUES_SUCCEEDED = 'FETCH_TOGGLE_VALUES_SUCCEEDED';
 
 export const SEARCH_QUERY_UPDATED = 'SEARCH_QUERY_UPDATED';
 export const CLEAR_SEARCH_RESULTS = 'CLEAR_SEARCH_RESULTS';
