+++ src/applications/facility-locator/utils/selectors.js
@@ -1,4 +1,4 @@
-export const toggleValues = state => state.featureToggles || {};
-export const isProduction = state => toggleValues(state).production;
+import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
+
 export const facilityLocatorShowCommunityCares = state =>
   toggleValues(state).facilityLocatorShowCommunityCares;
