+++ src/platform/site-wide/feature-toggles/actionTypes.js
@@ -1,0 +1,2 @@
+export const FETCH_TOGGLE_VALUES_STARTED = 'FETCH_TOGGLE_VALUES_STARTED';
+export const FETCH_TOGGLE_VALUES_SUCCEEDED = 'FETCH_TOGGLE_VALUES_SUCCEEDED';
