+++ src/platform/site-wide/feature-toggles/reducers/index.js
@@ -1,7 +1,7 @@
 import {
   FETCH_TOGGLE_VALUES_STARTED,
   FETCH_TOGGLE_VALUES_SUCCEEDED,
-} from '../utils/actionTypes';
+} from '../actionTypes';
 
 const INITIAL_STATE = {};
 
@@ -10,7 +10,6 @@ export const FeatureToggleReducer = (state = INITIAL_STATE, action) => {
     case FETCH_TOGGLE_VALUES_STARTED:
       return {
         ...state,
-        ...action.payload,
         loading: true,
       };
     case FETCH_TOGGLE_VALUES_SUCCEEDED:
