+++ src/platform/site-wide/feature-toggles/selectors.js
@@ -1,0 +1,2 @@
+export const toggleValues = state => state.featureToggles || {};
+export const isProduction = state => toggleValues(state).production;
