+++ src/platform/utilities/feature-toggles/feature-toggle-query-list.json
@@ -1,0 +1,3 @@
+{
+    "featureToggleQueryList": ["facilityLocatorShowCommunityCares"]
+}
